use ArgumentError instead of SyntaxError
The documentation for SyntaxError [1] says:

> Raised when encountering Ruby code with an invalid syntax

This exception isn't a subclass of StandardError, so a standard rescue
block won't catch it. It sounds like this exception is intended to be
used by the ruby VM and I suspect users will find it more convenient if
a subclass of StandardError is raised.

[1] http://ruby-doc.org/core-2.4.1/SyntaxError.html